+++ tests/Areas/Grafana/UnitTests/WorkspaceListCommandTests.cs
@@ -4,7 +4,6 @@
 using System.CommandLine;
 using System.CommandLine.Parsing;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 using AzureMcp.Areas.Grafana.Commands.Workspace;
 using AzureMcp.Areas.Grafana.Models.Workspace;
 using AzureMcp.Areas.Grafana.Services;