+++ tests/Areas/KeyVault/UnitTests/Key/KeyCreateCommandTests.cs
@@ -13,7 +13,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
-using NSubstitute.ReturnsExtensions;
 using Xunit;
 
 namespace AzureMcp.Tests.Areas.KeyVault.UnitTests.Key;